@@ -121,7 +121,7 @@ $(function() {
   });
 
   // Show licence form thingy
-  $("select#dataset_license_licenseLabel").on("change",function(){
+  $("select.licenseLabel").on("change",function(){
     var val = $(this).val();
     if(val == "Bespoke licence") $("#license-statement").show();
     else $("#license-statement").hide();
@@ -326,6 +326,17 @@ $(function() {
     }
   });
 
+  // Limit multi-valued roles to 3
+  $(".creatorRole").change(function(event) {
+    if ($(this).val().length > 3) {
+     alert('You can only choose 3!');
+     $("option:selected",this).each(function (index) {
+       if (index > 2) {
+         $(this).prop("selected", false);
+       }
+      });
+    }
+  });
 
   /* -------------------------------------------------------------
    * Tracker Follow
@@ -405,7 +416,24 @@ $(function() {
     }
   }
 
+  function assignReviewer() {
+  // Show text box for reviewer to assign to
+  $("#workflows_entries_status input[type='radio']").on("change",function(){
+    var val = $(this).val();
+    if (val == "Assigned") {
+      $("#assign-reviewer").css("display","block");
+      $("#workflows_entries_reviewer_id").removeAttr('disabled');
+    }
+    else {
+      $("#assign-reviewer").css("display","none");
+      $("#workflows_entries_reviewer_id").prop("disabled", true);
+    }
+  });
+  }
+
   fixTracker();
   expandPanel();
   $(document).on("scroll",fixTracker);
+  displayDoi();
+  assignReviewer();
 });
