@@ -15,7 +15,7 @@ class RegisterDoiJob
     begin
       dd.validate_required(payload)
     rescue ORA::DataValidationError => e
-      obj.workflowMetadata.update_status("System failure", e.message)
+      obj.workflowMetadata.update_status(Sufia.config.failure_status, e.message)
       obj.save!
       return
     end
@@ -23,16 +23,16 @@ class RegisterDoiJob
     begin
       dd.validate_xml(payload)
     rescue ORA::DataValidationError => e
-      obj.workflowMetadata.update_status("System failure", e.message)
+      obj.workflowMetadata.update_status(Sufia.config.failure_status, e.message)
       obj.save!
       return
     end
     # Register doi
     dd.call(payload)
     if dd.status
-      obj.workflowMetadata.update_status("DOI registered", dd.msg)
+      obj.workflowMetadata.update_status(Sufia.config.doi_status, dd.msg)
     else
-      obj.workflowMetadata.update_status("System failure", dd.msg)
+      obj.workflowMetadata.update_status(Sufia.config.failure_status, dd.msg)
     end
     obj.save!
   end
