@@ -23,7 +23,7 @@ module DoiMethods
     status = false
     if self.workflows.first && self.workflows.first.involves.any?
       self.workflows.first.involves.each do |event|
-        if event.include?("Register doi")
+        if event.include?(Sufia.config.doi_event)
           status = true
         end
       end
@@ -34,11 +34,32 @@ module DoiMethods
   def doi_registered?
     status = false
     if self.workflows.first && self.workflows.first.all_statuses
-      if self.workflows.first.all_statuses.include?("DOI registered")
+      if self.workflows.first.all_statuses.include?(Sufia.config.doi_status)
         status = true
       end
     end
     status
+  end
+
+  def request_doi
+    doi_s = self.doi(mint=false)
+    unless doi_s.blank?
+      return doi_s
+    end
+    doi_s = self.doi(mint=true)
+    unless doi_s
+      return doi_s
+    end
+    if self.publication.blank?
+      args = {'id' => "info:fedora/%s#publicationActivity" % self.id, :type => RDF::PROV.Activity}
+      self.publication.build(args)
+    end
+    if self.publication[0].hasDocument.blank?
+      args = {'id' => "info:fedora/%s#publicationDocument" % self.id}
+      self.publication[0].hasDocument.build(args)
+    end
+    self.publication[0].hasDocument[0].doi = doi_s
+    doi_s
   end
 
   def doi_data
@@ -111,14 +132,15 @@ module DoiMethods
              !cr.agent.first.affiliation.first.name.first.blank?)
             c[:affiliation] = cr.agent.first.affiliation.first.name.first
           end
-          if cr.role.first.to_s == "http://purl.org/dc/terms/creator"
+          if cr.role.include?(RDF::DC.creator)
             doi_data[:creator] << c
           else
-            if (Sufia.config.role_labels.include?(cr.role.first) &&
-               contributorTypes.include?(Sufia.config.role_labels[cr.role.first]))
-              c[:type] = Sufia.config.role_labels[cr.role.first]
-            elsif contributorTypes.include?(cr.role.first)
-              c[:type] = cr.role.first
+            matching_roles1 = cr.role.map { |role| Sufia.config.role_labels[role] if Sufia.config.role_labels.include?(role) && contributorTypes.include?(Sufia.config.role_labels[role]) }
+            matching_roles2 = cr.role.select { |role| contributorTypes.include?(role) }
+            if matching_roles1
+              c[:type] = matching_roles1.first
+            elsif matching_roles2
+              c[:type] = matching_roles2.first
             else
               c[:type] = "Other"
             end
