@@ -19,7 +19,7 @@ require 'fields/publication_activity'
 class ArticleRdfDatastream < ActiveFedora::NtriplesRDFDatastream
   #include ModelHelper
 
-  attr_accessor :title, :subtitle, :abstract, :subject, :keyword, :worktype, :medium, :language, :publicationStatus, :reviewStatus, :license, :dateCopyrighted, :rightsHolder, :rightsHolderGroup, :rights, :rightsActivity, :creation, :funding, :publication
+  attr_accessor :title, :subtitle, :abstract, :subject, :keyword, :worktype, :medium, :language, :publicationStatus, :reviewStatus, :license, :dateCopyrighted, :rightsHolder, :rights, :rightsActivity, :creation, :funding, :publication
 
   rdf_type rdf_type RDF::PROV.Entity
   map_predicates do |map|
@@ -49,7 +49,6 @@ class ArticleRdfDatastream < ActiveFedora::NtriplesRDFDatastream
     map.license(:in => RDF::DC, class_name:"LicenseStatement")
     map.dateCopyrighted(:in => RDF::DC)
     map.rightsHolder(:in => RDF::DC)
-    map.rightsHolderGroup(:in => RDF::ORA)
     map.rights(:in => RDF::DC, class_name:"RightsStatement")
     map.rightsActivity(:in => RDF::PROV, :to => "hadActivity", class_name:"RightsActivity")
     # -- creation activity --
@@ -88,7 +87,6 @@ class ArticleRdfDatastream < ActiveFedora::NtriplesRDFDatastream
     # Need to validate data and convert it to proper date format before indexing as date
     solr_doc[Solrizer.solr_name("desc_metadata__dateCopyrighted", :stored_searchable)] = self.dateCopyrighted
     solr_doc[Solrizer.solr_name("desc_metadata__rightsHolder", :stored_searchable)] = self.rightsHolder
-    solr_doc[Solrizer.solr_name("desc_metadata__rightsHolderGroup", :stored_searchable)] = self.rightsHolderGroup
     # Index the type of work
     self.worktype.each do |w|
       already_indexed = []
