@@ -11,9 +11,13 @@ class User < ActiveRecord::Base
   # Include default devise modules. Others available are:
   # :token_authenticatable, :confirmable,
   # :lockable, :timeoutable and :omniauthable
-  #devise :remote_user_authenticatable, :database_authenticatable, :registerable,
-  devise :database_authenticatable, :registerable,
+  if Rails.env.production?
+    devise :remote_user_authenticatable, :database_authenticatable, :registerable,
          :recoverable, :rememberable, :trackable, :validatable
+  else
+    devise :database_authenticatable, :registerable,
+         :recoverable, :rememberable, :trackable, :validatable
+  end
 
   # Setup accessible (or protected) attributes for your model
   #attr_accessible :email, :password, :password_confirmation, :remember_me
@@ -26,32 +30,37 @@ class User < ActiveRecord::Base
     email
   end
 
-  #def user_info
-  #  c = Qa::Authorities::Cud.new
-  #  ans = c.search(user_key, 'sso_username_exact')
-  #  if !ans.nil? && !ans[0].nil? && ans[0]['sso_username'] == user_key
-  #    return ans[0]
-  #  end
-  #  return nil
-  #end
+  def user_info
+    if Rails.env.production?
+      c = Qa::Authorities::Cud.new
+      ans = c.search(user_key, 'sso_username_exact')
+      if !ans.nil? && !ans[0].nil? && ans[0]['sso_username'] == user_key
+        return ans[0]
+      end
+    end
+    return nil
+  end
 
-  #def display_name
-  #  if self.user_info
-  #    if self.user_info['firstname'] and self.user_info['lastname']
-  #      return "#{self.user_info['firstname']} #{self.user_info['lastname']}"
-  #    elsif self.user_info['lastname']
-  #      return "#{self.user_info['lastname']}"
-  #    elsif self.user_info['firstname']
-  #      return "#{self.user_info['firstname']}"
-  #    end
-  #  end
-  #  return nil
-  #end
+  def display_name(user_info=nil)
+    # Passing in user_info as a parameter rather than calling self.user_info to minimie calls to CUD.
+    # NOTE: consequence of this is that sufia user model will display user key as name
+    if user_info
+      if user_info.fetch('firstname', nil) and user_info.fetch('lastname', nil)
+        return "#{user_info['firstname']} #{user_info['lastname']}"
+      elsif user_info.fetch('lastname', nil)
+        return "#{user_info['lastname']}"
+      elsif user_info.fetch('firstname', nil)
+        return "#{user_info['firstname']}"
+      end
+    end
+    return nil
+  end
 
-  def oxford_email
-    #if self.user_info
-    #  return self.user_info['oxford_email']
-    #end
+  def oxford_email(user_info=nil)
+    # Passing in user_info as a parameter rather than calling self.user_info to minimie calls to CUD.
+    if user_info and user_info.fetch('oxford_email', nil)
+      return user_info['oxford_email']
+    end
     return nil
   end
  
