@@ -9,51 +9,82 @@ class WorkflowPublisher
   end
   
   def perform_action(current_user)
+    wf_id = 'MediatedSubmission'
+    # Mint a doi or check a DOI if doi_requested? and in review status
+    mint_and_check_doi
     # send email
-    models = { "Article" => 'articles', "DatasetAgreement" => "dataset_agreements", "Dataset" => "datasets" }
-    record_url = Rails.application.routes.url_helpers.url_for(:controller => models[@parent_model.model_klass], :action=>'show', :id => @parent_model.id)
-    data = {"record_id" => @parent_model.id, "record_url" => record_url, "doi_requested"=>@parent_model.doi_requested?}
-    if @parent_model.doi_requested?
-      data["doi"] = @parent_model.doi(mint=false)
-    end
-    ans = @parent_model.datastreams["workflowMetadata"].send_email("MediatedSubmission", data, current_user, @parent_model.model_klass)
+    send_email(wf_id, current_user)
     # publish record
-    publish_record("MediatedSubmission", current_user)
+    publish_record(wf_id, current_user)
   end
 
   private
 
+  def mint_and_check_doi(wf_id='MediatedSubmission')
+    status = true
+    msg = []
+    # Check if in a review status
+    wf = @parent_model.workflows.select{|wf| wf.identifier.first == wf_id}.first
+    unless Sufia.config.user_edit_status.include?(wf.current_status)
+      # Mint a doi or check a DOI, if doi_requested?
+      if @parent_model.model_klass == 'Dataset' && @parent_model.doi_requested? && !@parent_model.doi_registered?
+        doi = @parent_model.doi(mint=false)
+        if doi.blank?
+          doi = @parent_model.request_doi
+          if doi
+            msg << "DOI '#{doi}' has been minted to register"
+          else
+            msg = 'Error minting a doi'
+            status = false
+          end
+        elsif doi.start_with?(Sufia.config.doi_credentials['shoulder'])
+          msg << "Using given doi '#{doi}' to register DOI"
+        else
+          msg << "Error: The DOI '#{doi}' is not a Bodleian DOI. It will not be registered!"
+          status = false
+        end
+      end
+    end
+    unless status
+      @parent_model.workflowMetadata.update_status(Sufia.config.failure_status, msg)
+    end
+  end
+
+  def send_email(wf_id, current_user)
+    models = { 'Article' => 'articles', 'DatasetAgreement' => 'dataset_agreements', 'Dataset' => 'datasets' }
+    record_url = Rails.application.routes.url_helpers.url_for(:controller => models[@parent_model.model_klass], :action=>'show', :id => @parent_model.id)
+    data = {'record_id' => @parent_model.id, 'record_url' => record_url, 'doi_requested'=>@parent_model.doi_requested?}
+    if @parent_model.doi_requested? && !@parent_model.doi_registered?
+      data['doi'] = @parent_model.doi(mint=false)
+    end
+    @parent_model.datastreams['workflowMetadata'].send_email(wf_id, data, current_user, @parent_model.model_klass)
+  end
+
   def publish_record(wf_id, current_user)
     # Send pid and list of open datastreams to queue
     # If datastreams are empty, that means record is all dark
+    #1 Check if ready to publish
     unless ready_to_publish?(wf_id=wf_id)
       return
     end
-    ans, msg = check_minimum_metadata
+    msg = []
+    status = nil
+    #2 check minimum metadata
+    ans, msg2 = check_minimum_metadata
+    msg = msg + msg2
     unless ans
-      status = "System failure"
-    else
-      status = "System verified"
-      open_access_content = @parent_model.list_open_access_content
-      numberOfFiles = (open_access_content.select { |key| key.start_with?('content') }).length
-      msg << "Open access datastreams: %s."%open_access_content.join(", ")
-      if @parent_model.model_klass == "Dataset"
-        Resque.enqueue(DatabankPublishRecordJob, @parent_model.id.to_s, open_access_content, @parent_model.model_klass, numberOfFiles.to_s)
-      else
-        # Add to ora publish queue
-        args = {
-            'pid' => @parent_model.id.to_s,
-            'datastreams' => open_access_content,
-            'model' => @parent_model.model_klass,
-            'numberOfFiles' => numberOfFiles.to_s
-        }
-        Resque.redis.rpush(Sufia.config.ora_publish_queue_name, args.to_json)
-      end
+      status = Sufia.config.failure_status
+    end
+    #3 Add record to publish workflow
+    unless status == Sufia.config.failure_status
+      ans, msg2 = add_to_queue
+      msg = msg + msg2
+      status = Sufia.config.verified_status
     end
     @parent_model.workflowMetadata.update_status(status, msg)
   end
 
-  def ready_to_publish?(wf_id="MediatedSubmission")
+  def ready_to_publish?(wf_id='MediatedSubmission')
     wf = @parent_model.workflows.select{|wf| wf.identifier.first == wf_id}.first
     status = false
     if wf.nil?
@@ -67,7 +98,7 @@ class WorkflowPublisher
     end
     occurences = wf.all_statuses.select{|s| s == wf.current_status}
     occurence = Sufia.config.publish_to_queue_options[@parent_model.model_klass.downcase][wf.current_status]['occurence']
-    return (occurences.length == occurence) || occurence == "all"
+    return (occurences.length == occurence) || occurence == 'all'
   end
 
   def check_minimum_metadata
@@ -103,6 +134,27 @@ class WorkflowPublisher
           msg << e.message
         end
       end
+    end
+    return status, msg
+  end
+
+  def add_to_queue
+    msg = []
+    status = true
+    open_access_content = @parent_model.list_open_access_content
+    numberOfFiles = (open_access_content.select { |key| key.start_with?('content') }).length
+    msg << "Open access datastreams: %s."%open_access_content.join(', ')
+    if @parent_model.model_klass == 'Dataset'
+      Resque.enqueue(DatabankPublishRecordJob, @parent_model.id.to_s, open_access_content, @parent_model.model_klass, numberOfFiles.to_s)
+    else
+      # Add to ora publish queue
+      args = {
+        'pid' => @parent_model.id.to_s,
+        'datastreams' => open_access_content,
+        'model' => @parent_model.model_klass,
+        'numberOfFiles' => numberOfFiles.to_s
+      }
+      Resque.redis.rpush(Sufia.config.ora_publish_queue_name, args.to_json)
     end
     return status, msg
   end
