@@ -117,9 +117,9 @@ module ORA
     def update_status
       #Update the workflow status based on the outcome of create and uploads to Databank
       if self.status
-        @obj.workflowMetadata.update_status('Data migrated', self.msg)
+        @obj.workflowMetadata.update_status(Sufia.config.migrated_status, self.msg)
       else
-        @obj.workflowMetadata.update_status('System failure', self.msg)
+        @obj.workflowMetadata.update_status(Sufia.config.failure_status, self.msg)
       end
     end
   
@@ -127,16 +127,14 @@ module ORA
       # Update the content datastreams in the dataset object with the new Databank location of the content files and delete local copy of content file
       @content_files.each do |ds, fp|
         filename = File.basename fp
-        opts = @obj.datastream_opts(ds)
-        old_location = opts["dsLocation"]
         # Update file location in datastream
-        opts["dsLocation"] = {
+        location = {
           'silo' => self.silo,
           'dataset' => self.dataset,
           'filename' => filename,
           'url' => @databank.getUrl(self.silo, dataset=self.dataset, filename=filename)
         }
-        @obj.datastreams[ds].content = opts.to_json
+        @obj.update_datastream_location(ds, location)
       end
     end
 
