@@ -26,7 +26,7 @@ describe ArticlesController do
   describe '#show' do
     let(:article) do
       Article.create do |a|
-        params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  'http://vocab.ox.ac.uk/ora#author', 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'workflows_attributes'  =>  {'0'  =>  {'id'  =>  '_:g70336805082780', 'involves'  =>  '', 'entries_attributes'  =>  {'0'  =>  {'description'  =>  ''}}, 'comments_attributes'  =>  {'0'  =>  {'description'  =>  ''}}}}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
+        params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  ['http://vocab.ox.ac.uk/ora#author'], 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'workflows_attributes'  =>  {'0'  =>  {'id'  =>  '_:g70336805082780', 'involves'  =>  '', 'entries_attributes'  =>  {'0'  =>  {'description'  =>  ''}}, 'comments_attributes'  =>  {'0'  =>  {'description'  =>  ''}}}}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
         a.apply_permissions(@user)
         MetadataBuilder.new(a).build(params, [], @user.user_key)
       end
@@ -82,7 +82,7 @@ describe ArticlesController do
       article.delete
     end
     it 'creates a new article' do
-      params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  'http://vocab.ox.ac.uk/ora#author', 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'workflows_attributes'  =>  {'0'  =>  {'id'  =>  '_:g70336805082780', 'involves'  =>  '', 'entries_attributes'  =>  {'0'  =>  {'description'  =>  ''}}, 'comments_attributes'  =>  {'0'  =>  {'description'  =>  ''}}}}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
+      params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  ['http://vocab.ox.ac.uk/ora#author', 'http://vocab.ox.ac.uk/ora#copyrightHolder'], 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'workflows_attributes'  =>  {'0'  =>  {'id'  =>  '_:g70336805082780', 'involves'  =>  '', 'entries_attributes'  =>  {'0'  =>  {'description'  =>  ''}}, 'comments_attributes'  =>  {'0'  =>  {'description'  =>  ''}}}}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
       sign_in @user
 
       expect {
@@ -95,7 +95,7 @@ describe ArticlesController do
   describe '#edit' do
     let(:article) do
       Article.create do |a|
-        params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  'http://vocab.ox.ac.uk/ora#author', 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
+        params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  ['http://vocab.ox.ac.uk/ora#author'], 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
         a.apply_permissions(@user)
         MetadataBuilder.new(a).build(params, [], @user.user_key)
       end
@@ -119,7 +119,7 @@ describe ArticlesController do
   describe '#edit_detailed' do
     let(:article) do
       Article.create do |a|
-        params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  'http://vocab.ox.ac.uk/ora#author', 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
+        params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  ['http://vocab.ox.ac.uk/ora#author', 'http://vocab.ox.ac.uk/ora#copyrightHolder'], 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
         a.apply_permissions(@user)
         MetadataBuilder.new(a).build(params, [], @user.user_key)
       end
@@ -151,7 +151,7 @@ describe ArticlesController do
       article.delete
     end
 
-    it 'creates a new article' do
+    it 'updates the article' do
       params = {'title' => 'Article title'}
       sign_in @user
       expect(article.title).to eq([])
@@ -165,7 +165,7 @@ describe ArticlesController do
   describe '#destroy' do
     let(:article) do
       Article.create do |a|
-        params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  'http://vocab.ox.ac.uk/ora#author', 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
+        params = {'worktype'  =>  {'typeLabel'  =>  'Journal article'}, 'title'  =>  'test.docx', 'subtitle'  =>  '', 'abstract'  =>  '', 'publication'  =>  {'publicationStatus'  =>  '', 'reviewStatus'  =>  '', 'publisher_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  '', 'website'  =>  ''}}}}, 'dateAccepted'  =>  '', 'datePublished'  =>  '', 'location'  =>  '', 'hasDocument_attributes'  =>  {'0'  =>  {'doi'  =>  '', 'uri'  =>  '', 'identifier'  =>  '', 'series_attributes'  =>  {'0'  =>  {'title'  =>  ''}}, 'journal_attributes'  =>  {'0'  =>  {'title'  =>  '', 'issn'  =>  '', 'eissn'  =>  '', 'volume'  =>  '', 'issue'  =>  '', 'pages'  =>  ''}}}}}, 'subject'  =>  {'0'  =>  {'subjectLabel'  =>  '', 'subjectAuthority'  =>  '', 'subjectScheme'  =>  ''}}, 'keyword'  =>  [''], 'language'  =>  {'languageLabel'  =>  '', 'languageCode'  =>  '', 'languageAuthority'  =>  '', 'languageScheme'  =>  ''}, 'creation'  =>  {'creator_attributes'  =>  {'0'  =>  {'name'  =>  'Test Two', 'email'  =>  '', 'sameAs'  =>  '', 'role'  =>  ['http://vocab.ox.ac.uk/ora#author'], 'affiliation'  =>  {'name'  =>  '', 'sameAs'  =>  ''}}}}, 'qualifiedRelation'  =>  {'0'  =>  {'entity_attributes'  =>  {'0'  =>  {'title'  =>  '', 'description'  =>  '', 'identifier'  =>  '', 'citation'  =>  ''}}, 'relation'  =>  ''}}, 'oaStatus'  =>  '', 'oaReason'  =>  '', 'refException'  =>  '', 'hasPart'  =>  {'0'  =>  {'type'  =>  'Content', 'identifier'  =>  'content01', 'description'  =>  '', 'accessRights'  =>  {'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}}}, 'accessRights'  =>  {'embargoStatus'  =>  'Open access', 'embargoDate'  =>  {'end'  =>  {'label'  =>  '', 'date'  =>  ''}, 'duration'  =>  {'years'  =>  '', 'months'  =>  ''}, 'start'  =>  {'label'  =>  'Today', 'date'  =>  ''}}, 'embargoRelease'  =>  ''}, 'funding'  =>  {'funder_attributes'  =>  {'0'  =>  {'agent_attributes'  =>  {'0'  =>  {'name'  =>  ''}}, 'awards_attributes'  =>  {'0'  =>  {'grantNumber'  =>  ''}}, 'funds'  =>  ''}}}, 'license'  =>  {'licenseLabel'  =>  '', 'licenseStatement'  =>  ''}, 'permissions_attributes'  =>  {'0'  =>  {'type'  =>  'user', 'name'  =>  '', 'access'  =>  ''}}}.with_indifferent_access
         a.apply_permissions(@user)
         MetadataBuilder.new(a).build(params, [], @user.user_key)
       end
